+++ src/stories/containers/cu-about/cu-about-container.tsx
@@ -3,8 +3,8 @@ import styled from '@emotion/styled';
 import { Divider, Typography } from '@mui/material';
 import { useSelector } from 'react-redux';
 import { useNavigate, useParams, useSearchParams } from 'react-router-dom';
-import { getMarkdownInformation } from '../../../core/business-logic/core-unit-about';
-import { getFTEsFromCoreUnit, getRelateMipObjectFromCoreUnit } from '../../../core/business-logic/core-units';
+import { getMarkdownInformation, getRelateMipObjectFromCoreUnit } from '../../../core/business-logic/core-unit-about';
+import { getFTEsFromCoreUnit } from '../../../core/business-logic/core-units';
 import { useAppDispatch } from '../../../core/hooks/hooks';
 import { RootState } from '../../../core/store/store';
 import { filterData, getArrayParam, getStringParam } from '../../../core/utils/filters';
